+++ src/main/java/com/whydigit/efit/repo/CustomersRepo.java
@@ -11,7 +11,7 @@ import com.whydigit.efit.entity.CustomersVO;
 
 
 @Repository
-public interface CustomersRepo extends JpaRepository<CustomersVO, Integer> {
+public interface CustomersRepo extends JpaRepository<CustomersVO, Long> {
 	@Query(value = "select a from CustomersVO a Where a.orgId=?1")
 	List<CustomersVO> getAllCustomersByOrgId(Long orgId);
